Create README.md
This project attempts to analyze differences in counties with regard to the number of solar (PV) installations. The idea would be to help a solar installation company understand where they might want to expand their business or where it might be worthwhile to try to create new markets (perhaps by lobbying for renewable energy incentives). This GitHub page does not include the full analysis; it only contains the files necessary to present findings in the slides (available at https://sarahu3.github.io/US-Solar/#/) and to show how the data were compiled. 


Files:
- index.ipynb is the Jupyter Notebook used to create the slides (converted to index.html)
- Data_munging.ipynb shows the steps to combine the data from various sources. Where possible, data were automatically downloaded using Python scripts, but many of the datasets were manually downloaded.